requirements.txt
fastapi
firebase_admin
numpy
opencv_python
Pillow
protobuf
PyYAML
torch
ultralytics
uvicorn
timm
datetime